+++ bin/CHANGELOG.md
@@ -1,3 +1,7 @@
+# <unreleased>
+
+ * Moved the `console` mod stuff out into separate crate
+
 # 1.0.3 (October 17th, 2018)
 
  * Fixed typo in output ("index" → "arg" for `LongWithUnexpectedData`)